+++ tools/scripts/tools/8dir/get_8_dir_res_get_data.py
@@ -10,13 +10,6 @@ import sys
 
 from get_file import Get_file
 
-def median(data):
-    print(data)
-    data.sort()
-    half = len(data) // 2
-    return (data[half] + data[~half])/2
-
-error = 0.003
 
 class Get8Dir():
     def __init__(self) -> None:
@@ -26,6 +19,7 @@ class Get8Dir():
         self.time = None
         self.res = {}
         self.new_res = {}
+        self.measure_data = {}
         self.cvs =None
         self.init()
 
@@ -40,10 +34,6 @@ class Get8Dir():
             if d.startswith("cali"):
                 self.file = d
                 print(d)
-                return True
-        print("No calibration data file found...")
-        print("Script end...")
-        exit(0)  
 
     def get_data(self):
         print("Using readlines()")
@@ -52,6 +42,7 @@ class Get8Dir():
             # for line in Lines:
             #     count += 1
             #     print("Line{}: {}".format(count, line.strip()))
+  
 
     def get_device(self):
         for l in self.data:
@@ -77,8 +68,12 @@ class Get8Dir():
             print(su)
             print(self.get_error(i))
             self.res[su] = self.get_error(i)
+            self.measure_data[su] = self.get_measure_data(i)
+            print("``````````````````")
+            # print(self.measure_data)
 
     def get_successed(self,i):
+        # return the measure radian...
         d1 = re.search("Registering last succeeded result as: \[(-?\d+)(\.\d+), (-?\d+)(\.\d+), (-?\d+)(\.\d+)\]", self.data[i], re.M|re.I)
         if d1 is None:
             return None
@@ -86,26 +81,35 @@ class Get8Dir():
         d3 = re.search("Calibration is done successfully!", self.data[i+4], re.M|re.I)
         if d2 is None and d3 is None:
             return None
+        # return the measure radian...
         return float(d1.group(3)+d1.group(4))
 
     def get_error(self,i):
+        # return the calibration error..
         for index in range(20):
-            if "beacause the error is" not in self.data[index+i]:
-                continue
             d1 = re.search("beacause the error is: (-?\d+)(\.\d+)?m",self.data[index+i],re.M|re.I)
-            d_scientific_notation = re.search("beacause the error is: ([+|-]?\d\.\d+[Ee][-]?\d.)m",self.data[index+i],re.M|re.I)
-            if d_scientific_notation is not None:
-                print("Got scientific notation..{}".format(d_scientific_notation.group(1)))
-                # print(d_scientific_notation.group(1))
-                return float(d_scientific_notation.group(1))
             if d1 is None:
                 continue
             # print(d1.group())
             return float(d1.group(1)+d1.group(2))
         return None
 
+    def get_measure_data(self,i):
+        data = []
+        l = 0
+        while True:
+            if len(data) == 6:
+                # print(data)
+                data.reverse()
+                return data
+            d1 = re.search("Registering last succeeded result as: \[(-?\d+)(\.\d+)?, (-?\d+)(\.\d+)?, (-?\d+)(\.\d+)?\]",self.data[i-l],re.M|re.I)
+            if d1 is not None:
+                data.append([float(d1.group(1)+d1.group(2)), float(d1.group(3)+d1.group(4)), float(d1.group(5)+d1.group(6))])
+            l += 1
+
+
     def handle_res(self):
-        print(self.res)
+        # print(self.res)
         last_rad = 99.
         for k, v in self.res.items():
             if not math.isclose(last_rad, k, rel_tol=0.1):
@@ -114,7 +118,7 @@ class Get8Dir():
                 last_rad = k
             else:
                 self.new_res[last_rad].append(v)
-        print(self.new_res)
+        # print(self.new_res)
 
 
     def convert_cvs(self):
@@ -128,17 +132,10 @@ class Get8Dir():
             for k, v in self.new_res.items():
                 data = []
                 data.append(k)
-                mean_of_data = median(v)
-                print("this dir data: {}".format(v))
-                print("mean of data: {}".format(mean_of_data))
                 for d in range(10):
                     # if v[d] is not None:
                     try:
-                        if math.isclose(v[d], mean_of_data, abs_tol=error):
-                            data.append(v[d])
-                        else:
-                            print("data {} is outlier..".format(v[d]))
-                            data.append("")
+                        data.append(v[d])
                     except IndexError:
                         data.append("")
                 data.append("=AVERAGE(B"+str(i)+":K"+str(i)+")")
@@ -156,24 +153,18 @@ class Get8Dir():
         os.rename(self.file, new_file_name)
         shutil.copy(new_file_name,has_handle)
         os.remove(self.cvs)
-        os.remove(new_file_name)
 
 
 
 
 if __name__ == "__main__":
-    print("Note: this script has tested passed on v3.7.1...")
-    print("If something wrong on other version, please connect shenzhen office..")
-    args = sys.argv
-    print(args)
-    if len(args) == 1:
-        pass
-    else:
-        hostname = args[1] + ".local"
-        print("Getting file from .... {}".format(hostname))
-        gf = Get_file(hostname)
-        fn = gf.get_calibration_file_name()
-        gf.get_file(fn)
+    # args = sys.argv
+    # print(args)
+    # hostname = args[1] + ".local"
+    # print("Getting file from .... {}".format(hostname))
+    # gf = Get_file(hostname)
+    # fn = gf.get_calibration_file_name()
+    # gf.get_file(fn)
     gd = Get8Dir()
     gd.get_res()
     gd.handle_res()